Also remove pants_runtime_python_version from pants.ini
Until 1.15.0, we shouldn't have this here as it's an example new users might refer to, and we don't want them to try putting this value then using a version < 1.15.0.dev4 and being frustrated it doesn't work.

We don't even need it here because our ci.py script dynamically rewrites pants.ini to add this value during tests.