Install Python 3.7 on Xenial and all OSX CI shards (#37)
### Problem
Because we release `pantsbuild.pants` as a `cp36-abi3` wheel, we would expect it to work with Python 3.6+ once we land https://github.com/pantsbuild/setup/pull/32, so, we should test with 3.7 as a simple sanity check.

As prework for that, we need to ensure we have a Python 3.7 interpreter on every shard possible.

## Solution
Use a mix of pyenv and pre-installed Python versions to get 3.7 on every OSX shard and on Ubuntu Xenial (16.04).

Note we cannot get it on Ubuntu Precise (12.04) or Trusty (14.04) without significant trouble, as their OpenSSL version is too outdated. Even updating OpenSSL with `apt-get` does not solve the issue. Because users still using Precise—which is EOL since April 2017—or Trusty—which is EOL starting April 2019—are unlikely to be using the most modern Python possible, we accept that we can't test 3.7  and only test 2.7 and 3.6 on those two shards.

## Result
All shards but Precise and Trusty now have Python 3.7, as proven with this test run: https://travis-ci.org/pantsbuild/setup/builds/506415674.

This does slow down our CI for the OSX shards, as they now must install Python 3.7 in addition to Python 3.6. They each go from taking ~5.5 minutes to 8 minutes. Because this repo's CI is so fast already and we don't run PRs against it often, this tradeoff is well worth the benefit we get through increased test coverage.

As a followup PR, we could introduce caching to the `PYENV_ROOT` for OSX shards and Ubuntu Precise.