Add shellcheck.py script and improve ./pants with results (#43)
### Problem
The entire point of this repo is the `./pants` script, so we should be using as much CI as possible to ensure it works well. While we do now have `ci.py` to act as an integration test, we would benefit from a linter.

### Solution
Use `shellcheck`, which is the most used and recommended linter for Bash scripts. 

Specifically:
* Add new `shellcheck.py` script
* Add new CI shard that runs `shellcheck.py` to avoid regressions
* Fix issues discovered in `./pants`