Extend ci.py to allow testing 1.14.0, 1.15.0, 1.16.0, and 1.17.0.
Each of these introduce significant churn in terms of which Python interpreters are supported. Now that the `./pants` script will set its default interpreter based on each of these pants_version, we should be testing each of them distinctly.

We also still test that `unspecified` works. This should be in the same .travis.yml entry as the most recent stable release.