For versions > 1.17.x, no longer assume Py36 support
We keep in in the list of possible interpreters, but go from max to min, i.e. 3.9->3.8->3.7->3.6. This is because we will always have a lower limit, but claim to not have an upper limit (beyond <4). If the user does not have 3.8 or 3.9 on their system, e.g. they aren't even released, we'll simply fall back to the prior version.

This reduces the risk of us dropping Py36 and then the `./pants` script being outdated for all our users as a result.