Determine default Python interpreter based on the pants_version being used (#49)
### Problem
Now that using Python 2.7 is ~deprecated and support will be removed soon, we need to stop defaulting to 2.7. 

Defaulting to Python 3 will make for a less stark transition. Specifically, users will be able to leave out `pants_runtime_python_version` from their `pants.ini`, which is good because we will deprecate this option once we drop Py2.

However, we do not want to break backwards compatibility. A user should be able to downgrade to 1.14.0 without issues.

### Solution
Determine which `pants_version` the user has pinned, and set the default interpreter from there, as follows:

* `<= 1.14.0`: Python 2.7
* `== 1.15.0`: Python 3.6->2.7
* `== 1.16.0`: Python 3.6->3.7->2.7
* `== 1.17.0`: Python 3.6->3.7->3.8->3.9
* `> 1.17.0`: Python 3.9->3.8->3.7->3.6

Notably, we first default to Python 3 and only fall back to Python 2 when necessary. We do this because we want most people using Python 3 so that we can completely stress test it before dropping Python 2.

For versions `> 1.17.0`, we default to using the max interpreter possible and only fall back to 3.6 as a last resort. This allows us to make no assumptions about Pants supporting 3.6 in the future.

If the user has not pinned `pants_version`, such as happens with first time users, we use the default interpreters for the current major release. With 1.15.0, this means we would use 3.6 or 2.7. Until 1.18.0 is released and our migration is stabilized, we will update the default case in the `./pants` script to match that new release. This should only impact new users before they run the `./pants generate-pants-ini` command.

We also avoid the ambiguous `python3` bin name to avoid accidentally trying to use Python 3.5, for example.