Stop using Pyenv shims for custom installed Python (#56)
Shims are designed for humans, not machines. Most tangibly, they resulted in this unexpected behavior https://github.com/pantsbuild/pants/issues/7636 on the OSX 10.12 shard.

We still use `pyenv global` on the Precise and Xenial shards, however, because Travis pre-installs Python via Pyenv on those shards and this is the standard way to fix an issue with their implementation so that the interpreters become discoverable on the `PATH`. We stick with Travis's pre-installed versions for performance and a smaller cache, at the cost of a less uniform and more complex Python setup across CI.