Replace `ci.py` with `test_sanity_check.py` for less code (#60)
## Problem
When writing `ci.py`, the focus was primarily on re-implementing the original YAML-based tests, along with testing all the new permutations of using Python 3. This got the job done, but resulted in a lot of code.

## Solution
Now that we have infrastructure in place for things like making a temporary folder with `./pants` copied into it, `ci.py` can be replaced by a far simpler test file.

## Result
We test the exact same thing, but with far less code.