Use Tox instead of `build-support` scripts (#64)
When we expanded our CI in May 2019, we took the approach of Pants of having multiple lint utility scripts in `build-support`, rather than Pex's approach of using `tox`. This was mostly due to my unfamiliarity with Tox.

Beyond unnecessary custom code, another downside to this `build-support` approach is that it poses a chicken-and-the-egg problem when developing this repo. For example, for our tests and linters to run properly, `./pants` must work; however, the entire point of our linters and tests is to confirm that the script `./pants` works in the first place. Using `tox` allows us to run our tests and lints regardless of the condition of the `./pants` script.

This also does the following:
* Adds to CI macOS Mojave and Ubuntu 18.04.
* Applies Black formatting
* Uses Flake8 rather than the custom Pants checker