Use pyenv to test Python interpreter resolution rather than `--pants-runtime-python-version` (#67)
We previously used `--pants-runtime-python-version` to force Pants to use certain Python versions. This was not safe to keep doing, though, because we ended up removing this option.

Instead, we can use Pyenv to get the same effect.

This will allow us to start testing default Python interpreter resolution for Pants > 1.16.

We remove Ubuntu 12.04 from CI to land this, as it has been EOL since April 2017 and causes issues with getting Pyenv to work properly.