Remove support for Pants <= 1.16.0 (#68)
Pants 1.16.0 is over 9 months old. For new Pants users, it is highly unlikely they will change the default version of Pants 1.24.0 all the way back to 1.16.0. 

By removing this support, we can guarantee that Pants is running with Python 3.6+. This allows us to simplify much of the `./pants` code, e.g. allowing us in a followup to use the stdlib `venv` implementation instead of `virtualenv`.

If users try using this script with Pants 1.16.0 or earlier, we point them to a revision that will work properly.