Convert tests to function style.
This moves tests from `TestCase` subclasses to test functions and pytest
fixtures. Benefits include generally more straight forward typing which
makes it easier to do the right thing for setup/teardown in terms of
fixture scope as well as eliminating most explicit contextmanager use;
thereby reducing nesting.

Pytest has a bit of a magic curve and this change does not document the
magic, but instead expects maintainers to read up a bit and learn about
fixtures dependency injection by name behavior and the pre-defined
fixtures supplied by pytest like `tmp_path` and `monkeypatch`.

This is a follow up to #67:
  https://github.com/pantsbuild/setup/pull/67#discussion_r382932317