Fix first_time test: make robust to warnings or other stderr (#78)
Currently, the `test_only_bootstraps_the_first_time` test is failing, as a result of brittleness to warnings and other textual output that is now present when pants runs.

It might be appropriate to update the configuration used when bootstrapping pants so that these warnings and messages are not present, but regardless I think updating the assertions here makes sense and is more future-proof.


Example of build failure:
https://travis-ci.com/github/pantsbuild/setup/jobs/339413866


Sidenote: it's unfortunate that this test broke due to changes in another repo. Might be a good argument for consolidating this setup code into the main pants repository.