Parameterize the config file name.
Instead of hardcoding pants.toml/pants.ini.

Also update a test to use pants 1.28.0. That test
no longer passed, even at unmodified HEAD, due to
some watchman issue. In any case, it's more important
to test against a recent pants, since people downloading
the script are likely to be doing so in order to run
the most recent stable release (which is what the setup
script will choose for them on first bootstrap).