Require `pants_version` to be set (#86)
We generally highly encourage users to use lockfiles because it's important for deterministic builds. But with the Pants script itself, we did not require this determinism, which could result in builds breaking randomly when we do new stable releases.

Our docs already have been saying to set the `pants_version` before running `./pants`, so we don't need to update anything there.