Pants 2.5+ supports Python 3.9 (#96)
This also defaults to Python 3.8 over 3.7 for Pants 2.2+, given the performance improvements.

This stops testing that we can package `pants_requirement()` when using a SHA, as this is the wrong place to test that. That belongs in a test in pantsbuild/pants. It was non-trivial to get the test to still work across 1.30 and 2.x due to issues like renaming `python_binary` to `pex_binary`.