Fix scrub msg in pants script and upgrade PEX version (#127)
I am opening this PR to modify the function `scrub_env_vars()` in the `pants` script so that it scrubs all `PEX_*` environment variables except `PEX_ROOT`. I have also upgraded the variables `_PEX_VERSION` and `_PEX_EXPECTED_SHA256` in the script but in a separate commit just in case we want to cherry pick and merge only the one related to scrubbing PEX_* environment variables.