Add 'bootstrap tools' via PANTS_BOOTSTRAP_TOOLS=1
This functionality that can introspect the bootstrapping process,
similar to `PEX_TOOLS=1 ./some.pex ...` being able to introspect a PEX
file.

The motivating tool, added here, is `PANTS_BOOTSTRAP_TOOLS=1 ./pants
bootstrap-cache-key` that prints a somewhat opaque string designed to
be used as a key for CI caching systems. This is designed to help with
https://github.com/pantsbuild/actions/issues/5.